Allow pg_auto_parameterize extension to use placeholder loaders
The most important benefit from doing so is to allow the eager
loading limited associations to use the union eager loader,
instead of falling back to request the entire result set and
then limiting it in Ruby.

This adds Dataset#placeholder_literalizer_class and
placeholder_literalizer_loader to DRY up code that creates and
uses the placeholder literalizers.

This changes the internal PlaceholderLiterlizer::Recorder API
to accept the PlaceholderLiterlizer as the first argument, to
allow for easier subclassing.  This is a backward incompatible
change for anyone calling PlaceholderLiterlizer::Recorder.loader
directly.

This adds PlaceholderLiterlizer#sql_origin, to allow easier
overriding of the starting SQL object (an empty string by default).

This adds PlaceholderLiterlizer#append_sql, to allow appending to
a given SQL string.  This modifies the union eager loader to use
that instead of assuming the SQL will all be plain strings, and
then joining them together.

This adds a subclass of PlaceholderLiterlizer to the pg_auto_parameterize
extension, which overrides sql_origin.  This adds support QueryString#dup
to the extension, to also duplicate the stored arguments.  Additionally,
when freezing the QueryString, the arg_map is also frozen, not just the
args.