Make pg_auto_parameterize placeholder loader actually speed things up
Before, the pg_auto_parameterize placeholder loader was slightly
slower.  This was because it was calculating the SQL the long way
in addition to the fast way on every call.  Store the empty base
SQL string when creating the loader (which includes the args), and
make a copy of it for every load.  This provides about a 4% speedup
for Dataset#first for a simple query, with larger speedups for
more complex queries.