Make Dataset#count with no argument/block handle dataset with custom SQL using ORDER BY on MSSQL
The spec commited for Dataset#empty? with custom SQL did not work
on MSSQL since it used an ordered query, and you cannot have an
ordered subquery on MSSQL unless it is limited.

Try to fix the #empty? issue by seeing if the dataset returns at
least one row.  For #count with no arguments or block, count all
rows of the dataset. For #count with arguments or block, there
is no way to ensure correct behavior, so use the default behavior
and hope the custom SQL is either not ordered or uses a limit or
offset.

Tweak empty? spec so that is passes on Derby.