Add bigdecimal as an explicit dependency
While sequel always requires bigdecimal, it's not listed as a
dependency. It's possible to have a Ruby installation without bigdecimal
(on EL8 / Fedora: `dnf install ruby --setopt=install_weak_deps=False`)
and this helps. While Fedora already explicitly adds it as a dependency,
listing it in the gemspec allows the dependency generator to pick it up
automatically.