Include :min_value and :max_value schema entries for date/timestamp/timestamptz columns on PostgreSQL 9.6+
This follows the additions of :min_value and :max_value support for
integer and numeric columns.

PostgreSQL <9.6 support is possible, and probably just requires
different min and max values.  If someone is interested in that,
they should send a pull request.

It's also possible to support this on other databases.  I'll
consider pull requests to support other databases.

The timestamp support is only included if Sequel.datetime_class
is Time.  There's not a good reason anymore to set it to DateTime.
If someone wants to support this for timestamps when
Sequel.datetime_class is DateTime, please submit a pull request.

This modifies the pg_extended_date_support extension to work
around a bug in Ruby 1.9's Date#<, which was necessary to get
the specs passing.