Fix recent PostgreSQL spec failure when using pg_auto_parameterize
The way pg_auto_parameterize works, pg_extended_date_support needs
to be loaded for the max/min value date/time/timestamp spec to run.
For timestamptz, an explicit cast is needed to pass.

Fixing it caused a later spec-order dependent bug, so skip that
specific assertion when testing with pg_auto_parameterize.