Make constraint_validations plugin not validate missing columns that have a default value
auto_validations already has similar code.  If a column is missing,
but has a default value, we can assume the default value is valid.

While here, add :from=>:values to the validation options, so the
validated value comes from the underlying column value and not
any wrapped value.  Seems unlikely a user would want constraint
validations on a column where the method returned a different value
from the underlying column, because then the database and model
checks would be checking different values.