Add validation_helpers_generic_type_messages plugin for more useful type validation failure messages (Fixes #2028)
I'd like to make this change directly to validation_helpers, but
I think it is too backwards incompatible to do outside a major
release, mostly because the current default messages are likely
used as the basis for translations.  So instead of modifying the
default behavior, allow to opt-in to it via another plugin,
validation_helpers_generic_type_messages. I expect that the
validation_helpers_generic_type_messages behavior will become the
default in Sequel 6.