Add trilogy adapter
Trilogy support was recently added to ActiveRecord, and it is looking
like it will be the primary mysql driver at some point in the future.
One nice feature is that it does not require any mysql client
libaries installed to install the gem.

As adapters go, this was very easy to write.  However, the following
are not supported well (most are the same issues as mysql2):

* time types (does not return Sequel::SQLTime)
* blob types (does not return Sequel::SQL::Blob)
* application timezone different than database timezone
* Sequel.datetime_class = DateTime

It's possible the above could be fixed with dataset extensions, which
can be considered on an as-requested basis (would slow down result
retrieval).

Additionally, trilogy does not support returning the correct number
of matched rows on UPDATE (it uses affected rows, skipping rows
where the values in UPDATE match the current row). I think trilogy
itself needs to be fixed to support that.