Make mysql adapter work with ruby-mysql 4+
Note that multi-results support does not appear to be working
when using ruby-mysql 4+, so skip the related specs.  That
doesn't work on the mysql2 or trilogy adapter either, and
is not supported in general with Sequel. I'll accept patches
to fix this, assuming it is possible in ruby-mysql 4.

Create a Sequel::MySQL::Adapter subclass of Mysql when using
ruby-mysql 3+ to avoid the need for each connection to have
a singleton class.

When ruby-mysql 3+ is used, force the :compress option to
false to avoid a warning.  This still warns when using
ruby-mysql 2, but I'll leave that for someone who is affected
by it and cares enough to fix it.