Add sharded_timed_queue connection pool
This adds a sharded version of the timed_queue connection pool, and makes the
connection_expiration, connection_validator, and server_block extension
work with it.

While working on this, I found a bug in the sharded_threaded connection
pool's remove_server method, so this fixes that.

To get better testing of this connection pool, this adds support for the
SEQUEL_DEFAULT_CONNECTION_POOL environment variable, which sets the
default :pool_class option if one is not specified.  Various changes
are made to the specs such that you can set this environment variable
when running the specs and have the specs still pass.

I plan to make the timed_queue and sharded_timed_queue connection
pools the default connection pools on Ruby 3.2 in the future.  It's
possible that will be done in Sequel 6, but it could be earlier
depending on feedback.