Add pg_xmin_optimistic_locking plugin for optimistic locking for all models without database changes
This implements similar support to the optimistic_locking plugin,
but without the need for database changes to add a lock version
column.  Instead, it uses the xmin system column, which is
automatically updated every time the row is updated.

This allows you to load the plugin into a base class and have
all models under the base class support optimistic locking.

Internally, this splits off an optimistic_locking_base internal
plugin from optimistic_locking, makes optimistic_locking and
mssql_optimistic_locking use it, and uses it as the basis
for the new pg_xmin_optimistic_locking plugin.