Use deep copy when returning default value
Previously, default values were returned as a shallow copy, which works fine
for most of the cases. However if the returned object is a complex object such
as Hash or Array, which can be representation of PostgreSQL's JSONB column for
example, shallow copy is not enough. If caller updates the returned object,
default value also gets overwritten, thus all subsequent calls use the updated
value.

Marshal.dump & Marshal.load is used to generate a deep copy of the object,
which might be an overkill, but this seems like a most reliable way of creating
a deep copy.