Add MassAssignmentRestriction#column
Update MassAssignmentRestriction.create to take the column. Also, have it always include
the class in the error message, even if the class is anonymous.

For consistency, the column is always saved as string, since the mass assignment methods
are generally used with string-keyed hashes provided by the user, and the comparison to
allowed methods is done using strings.

Simplify some of the related test code.