Add paged_update_delete plugin for paged deletes and updates
This allows for splitting a large update or delete into multiple
queries to prevent locking the related database table for a
long period of time.  Each update or delete query operates on a
range of primary key values.

Split a _force_primary_key_order private model dataset method
from _primary_key_order and use it in the plugin.