Replace paged_operations_size method with :rows_per_page option
I did not use this approach when developing the plugin because I
was using a splat for paged_update.  However, as Dataset#update
only takes a hash and not a splat, paged_update should do the same,
making it possible to support an options hash.  Currently, only the
:rows_per_page option is supported.

Minor documentation updates while here.