Avoid infinite loop when handling exceptions with a cause loop in jdbc adapter
Doubt any jdbc driver actually creates a cause loop, but having
bounded iteration seems like a better approach here.

Also, avoid checking exception causes if the exception is already
recognized as a disconnect.

This also fixes cases where the exception has a cause but the SQL
state is part of the exception and not its cause.

Remove the database_exception_sqlstate in the jdbc/db2 adapter,
since it should be handled by the generic jdbc code now, as
sql_state is probably calling the getSQLState Java method.