Add Database#{defer,immediate}_constraints on PostgreSQL for changing handling of deferrable constraints in a transaction
This allows you to easily check deferrable constraints at a
specific point in a transaction, instead of waiting until the
end of the transaction.  For deferrable constraints that are
initially immediate, this allows you to deter the checking of
the constraints.

I thought about using a single set_constraints method for this,
but I think two separate methods results in more readable code.