Support INSERT RETURNING on MariaDB 10.5+, and use it when saving new model objects
Do not use this on jdbc/mysql, because the MySQL JDBC driver
raises an exception before sending the query to the server because
it parses the query and because it is INSERT, and MySQL does
not support INSERT returning result sets.

This is probably fixable by using the MariaDB JDBC driver, but
Sequel doesn't yet support that.  The jdbc-mariadb Ruby gem is
old enough that it is unlikely to have support.  If the
jdbc-mariadb gem is updated to a version that supports this,
we can add a jdbc/mariadb adapter and support INSERT RETURNING
when using it.