Add auto_cast_date_and_time extension, for casting date and time values using SQL standard functions
Previously, Sequel::Dataset used #requires_sql_standard_datetimes? to
detemine whether to cast date and time values.  This was inefficient,
and there wasn't an easy way to turn this on for specific databases,
it was either always on or always off based on the database adapter
in use.

Instead of checking #requires_sql_standard_datetimes?, this moves the
related code to an extension.  The databases that previously used this,
h2, hsqldb, and oracle, now load the extension file and include the
module in the related dataset class/module. The extension can optionally
be loaded on at least PostgreSQL and MySQL, for potentially better
behavior. It does not work on SQLite, as SQLite does not have a separate
date and time type. It does not work on MSSQL, which does not support
the standard SQL date/time conversion casts.

This handles the casting by overriding the related
Dataset#literal_{date,datetime,sqltime,time}_append methods, leaving
the #literal_{date,datetime,sqltime,time} methods as returning the
uncasted value.  To get bound argument code to work,
the #literal_{date,datetime,sqltime,time} methods are called directly,
stripping off quotes if necessary.  This will be simplified in future
commits.

The pg_timestamptz extension integrates with the auto_cast_date_and_time
extension, automatically casting to TIMESTAMP WITH TIME ZONE instead
of plain TIMESTAMP.

For testing, you can set the SEQUEL_AUTO_CAST_DATE_TIME environment
variable when running the adapter or integration specs, and it will
run them with the auto_cast_date_and_time loaded into the Database.

This is not backwards compatible with external adapters that
override #requires_sql_standard_datetimes? and expected the
literalization code to respect it.  Such external adapters must be
updated to require 'sequel/extensions/auto_cast_date_and_time' and
include AutoCastDateAndTime in the appropriate dataset class/module.