Remove special handling of %z modifier in Dataset#default_timestamp_format
This modifier was removed by databases that don't support timezones in
timestamps, and then special cased because %z didn't work correctly on
Windows historically (probably in Ruby 1.8).  I think the Windows %z
issues are fixed on Ruby 1.9+, so there should be no reason to avoid it.

There are only two databases where Sequel supports timezones in timestamps
by default, PostgreSQL and Oracle. Sequel can optionally support timezones
in timestamps on SQLite, but it isn't the default because it does not work
with the SQLite date/time functions.  Oracle already overrides
default_timestamp_format, this makes both PostgreSQL and SQLite override
it and set the appropriate format (conditional in the case of SQLite).

As the remaining databases do not support timezones in timestamps, this
removes %z usage from Dataset#default_timestamp_format.

This makes Dataset#supports_timestamp_timezones? no longer used internally.
It remains available as it could be useful to check whether timezones in
timestamps are supported.