Minor refactor of sqlite adapter regexp function support
Support "cached" in addition to :cached for the caching support,
so this can be used in connection URLs via
setup_regexp_function=cached.

Simplify the regexp function by always using indirection to get
whether the regexp matches.  Probably a minor performance hit,
but the uncached Regexp case is already slow, and the cached mode
will still be much faster than the previous uncached case.

Fix the CHANGELOG, so that newest entries are at the top, and add
the issue id.