Support Database#with_advisory_lock on PostgreSQL, MySQL, and Microsoft SQL Server
This supports advisory (explicit) locking, using the
database-specific APIs.  To work on both servers, lock ids
should be integers in the signed 64-bit range.

By default, an AdvisoryLockError is raised if the lock cannot be
immediately acquired.  You can use the :wait option to wait until
the lock can be acquired, instead of raising.