Add Dataset#select_prepend for prepending to the current selected columns
In Ruby 1.8, hashes weren't ordered, so this method would not have made
sense. However, since Ruby 1.9, hashes are insertion ordered, and
it isn't uncommon to rely upon the order.  Sequel orders hashes by their
order in the returned values from the database, which should be the
same as the order of columns in the SELECT clause. Sequel has had a
dataset method for appending to the SELECT list basically forever,
this adds support for prepending to the SELECT list.

There are quite a few databases that support `SELECT *, column` but
not `SELECT column, *`, so when using select_prepend, if there isn't
an existing selection, Sequel always uses qualified table selections.
If anyone really cares, we can add a separate method so that
databases that support `SELECT column, *` can use it. Note that
using select_prepend after select_append, where the select_append
used `SELECT *, column`, may break on databases that do not support
selected columns before a bare wildcard.