Add temporarily_release_connection Database extension for multithreaded transactional testing
This allows one thread to start a transaction, and then release
the connection back for usage by the connection pool, so that
other threads can operate on the connection object safely inside
the transaction.  This requires the connection pool be limited
to a single connection, to ensure that the released connection
can be reacquired.  It's not perfect, because if the connection
is disconnected and removed from the pool while temporarily
released, there is no way to handle that situation correctly.