Revert "Add :min_wait_timeout option to sharded_threaded connection pool and default to 0.1, fixes stalling observed on JRuby"
This reverts commit 72f08cd4cb1186b60d937df283317b29107c4506.

This implements a better fix, by rechecking for an available
connection inside the sync block before waiting on the condition
variable.  Because the recheck is inside the sync block, it should
avoid the race condition (barring a bug in the condition variable
signal/wait code).

Add an acquire_available private method to both connection pools
to handle this logic.