Add stdio_logger extension for a minimal logger that Sequel::Database can use
This is mostly taken from the bin/sequel implementation, with
the main change that it should be more thread-safe by using
write instead of puts.

Change bin/sequel to use this extension instead of defining an
anonymous class.  Also have bin/sequel use sync for log files,
as that is what logger does.