Support schema-qualified table names when using create_table :temp with a Sequel::SQL::QualifiedIdentifier (Fixes #2185)
1fd397626413b394bf771ea69669efc7b91c92f7 stopped supporting this
due to problems when setting default_schema on PostgreSQL.  However,
just using the inspect version of Sequel::SQL::QualifiedIdentifier
as the table name is not something any user wants.  PostgreSQL does
not support schema-qualified temporary tables, and I'm not sure
if any other databases do, but it's better to issue the query with
a schema qualified table and have it raise an error, as opposed
to having it silently succeed with a bogus name.