Restore unescaping of database file paths in the sqlite and amalgalite adapters (Fixes #2201)
The 5.83.0 release notes mentioned that uri_to_options was changed
to handle conversion of URI options, but considering that I missed
fixing this issue for the sqlite and amalgalite adapters that ship
with Sequel, expecting external adapters to handle changes seems
unreasonable.

Add an options_from_uri method that calls uri_to_options and handles
the options unescaping, and have Sequel call that in the two places
that previously called uri_to_options directly.

As this is a regression from the previously release, bump version
to 5.83.1.