Switch default connection pool to timed_queue on Ruby 3.2+
The timed_queue connection pool was already the default on Ruby
3.4+.  However, recent discussions have shown that it is much
better than the threaded connection pool, at least in cases where
a sampling profiler is used.

This should be a drop-in replacement for most applications. However,
the timed_queue connection pool does not support the
available_connections and allocated accessors that the threaded
connection pool supports.  Those accessors are planned for removal
in Sequel 6, but are not formally deprecated yet.  Applications
that are using those accessors should either stop, or specify
pool_class: :threaded to continue using the threaded connection
pool.