Make Dataset#get and #first without argument not create intermediate datasets if receiver uses raw SQL
There's no reason for a dataset clone in this case, as the
limit and select settings will be ignored.

This also fixes use of Dataset#get without argument when using the
implicit_subquery extension. Previously, it would try to use a
subquery in an unhelpful way:

  DB['SELECT a FROM b WHERE c = 1'].extension(:implicit_subquery).get
  #  SELECT NULL AS v FROM (SELECT a FROM b WHERE c = 1) AS t1 LIMIT 1

Now, it runs the query directly and returns the first value.