Support the :disable_dqs Database option in the sqlite adapter to disable treating double quoted values as strings
As described by the SQLite documentation, treating double quoted
values as strings instead of identifiers is a misfeature.

Unfortunately, the sqlite3 gem implements this using a :strict
option, which is poorly named, as you would think it would have
something to do with STRICT tables and it does not. Considering
SQLite already supported STRICT for tables before the sqlite3
gem added support for the :strict option, I'm not sure why the
option name was selected.

Use :disable_dqs as the name of this option in Sequel, following
the SQLite naming of the config options.  This currently does
not default to true for backwards compatibility, but I expect to
change that in Sequel 6.