Add pg_schema_caching extension, for reloading OIDs for custom types when loading cached schema
While PostgreSQL uses the same OIDs for all built-in types, custom
types have OIDs that differ for each database, even for databases
built from the same migrations.  That means that schema caches
built from development or test databases, if loaded into a
Database object for the production database, will have incorrect
oids for custom types.

This avoids the issue by replacing custom oids with :custom in
dumped schema. When loading schema, a single query is done to get
the oids for each custom type in the dumped schema, and the column
schema hashes are then updated to set the correct oid.