Support column aliases with data types on PostgreSQL, useful for selecting from functions returning records
Previously, you had to use Sequel.lit to handle this type of
derived column list.  While this is useful when selecting from
any database functions that return records, in terms of working
with other parts of Sequel, this makes usage of to_recordset in
the pg_json extension simpler, as shown by the spec change.