Add subset_static_cache plugin for statically caching subsets of a model class
This is useful if the entire model class is not static, but specific
subsets of the model class are static.  It operates like the static_cache
plugin, but restricted to specific subsets.

Update the static_cache_cache plugin to handle the subset_static_cache
plugin in addition to the static_cache plugin.