Add temporary names for internally defined anonymous classes and modules on Ruby 3.3+
These should hopefully aid debugging and ease understanding.

I don't believe any of these calls are in performance-sensitive
code.  Most are called only during application startup.  In terms
of runtime calls, Dataset#with_extend may be the most common, but
in that case, you are allocating a new class, so setting the temporary
name shouldn't be that significant in comparison.  The temporary names
are provided inside a block, so the dynamically created strings are
not allocated on Ruby < 3.3.