Have Database#valid_connection? if queries are currently blocked
If checking a connection for validity, and a BlockedQuery exception is
raised, treat it as a valid connection.  You cannot check whether the
connection is valid without issuing a query, and if queries are blocked,
you need to assume it is valid or assume it is not.  Since it most cases
it will be valid, this assumes validity.

An alternative approach to this would be to use allow_queries to
implicitly allow queries to determine validatity, but since that issues
a query, it violates the expectation of a user using block_queries. Such
behavior would have to be opt-in, and I'll wait for someone to request it
before implementing it.