Improve coverage of pg_eager_any_typed_array extension
There is no need to check for the :join_table_db/:separate_query_per_table
options, since associations that set those options do not call into
this code.

Add a spec for cases where the column type cannot be found in the schema.

Add a :nocov: around the case statement, since we want to skip handling
unsupported associations, but all association types that ship with
Sequel and call into this code are already handled.