Support treat_string_list_as_untyped_array Database option in pg_auto_parameterize_in_array extension
This option uses an uncasted parameter for the array. This allows the
IN/NOT IN to = ANY/!= ALL conversion for to work for string arrays that
represent non-text database types.  This is especially helpful when
using uuid and enum types, which typically represented as Ruby strings.

I originally thought that this wouldn't be feasible as it would
cause too many issues.  That's because PostgreSQL does not support
uncasted arrays in this case when using the ARRAY syntax:

  uuid_column = ANY(ARRAY[$1]) # uuid value as parameter
  # PG::UndefinedFunction: ERROR:  operator does not exist: uuid = text

When I first started looking into auto parameterization (back in the
Sequel 3 timeframe), I came across this issue, which I thought applied
to all arrays when using parameters.  That's one reason that
pg_auto_parameterize_in_array has historically casted the array
parameters.

However, it turns out that PostgreSQL allows using an uncasted parameter
for the entire array:

  uuid_column = ANY($1) # uuid[] value as parameter
  # No error

I'm not sure why PostgreSQL behavior differs for these two examples.
However, using the uncasted array appears to work for every type where
the elements of the RHS array can be implicitly casted to the type of
the LHS.

While working on this, I attempted to remove the explicit casting for all
array types.  Unfortunately, while that works for almost all cases, there
are cases where removing the explicit cast from the parameter breaks things.
These are cases where the LHS side type needs to be implicitly converted to
the scaler type of the array parameter.  Example, where the parameter is
an array with numeric elements:

  # Explicit cast works (integer_column is implicitly casted to numeric)
  integer_column = ANY($1::numeric[])

  # This breaks if parameter has numeric values, since implicitly casts
  # the parameter to integer[]
  integer_column = ANY($1)

So I cannot change pg_auto_parameterize_in_array to use that approach.
It's possible it could be added as a separate extension if there were
user demand for it.  It's possible users may not want the LHS to be
implicitly casted, as doing so can break usage of indexes.

While here, speed up the pg_auto_parameterize_in_array by avoiding
Database#typecast_value calls at runtime.  This change slightly breaks
backwards compatibility, because you now need to set the appropriate
option before loading the extension.