Add `allow_by_default` option for `forbid_lazy_load` plugin
This one was conceived and discussed more thoroughly over in #2286. The
basic idea is that the `forbid_lazy_load` plugin disables lazy loading
on specific model instances, but also broadly on any models loaded
through an association. The latter can be detrimental when trying to cut
over to eager loading incrementally and where it's hard to get an entire
codebase converted all at once.

Here, the `forbid_lazy_load` plugin picks up a new `allow_by_default`
option. When it's set to true, lazy loading can still be forbidden
normally on an instance-by-instance basis, but it's not forbidden
through associations. This should make incremental migrations to eager
loading with `forbid_lazy_load` somewhat easier to do.