Fix possible concurrency issue in eager loading due to Method#hash iterating over a hash that can be inserted into concurrently (Fixes #2292)
There was previously a concurrency issue, because the eager_loader
was used as a hash key for a temporary hash.  The default eager
loader is a Method object from the AssociationReflection. As
AssociationReflection is a subclass of Hash, and Method#hash
calls #hash on the receiver of the Method, and Hash#hash iterates
over the hash, this meant that using the eager loader as a hash
key iterated over the hash.  Iterating over the hash itself is
not a problem, as the hash is not modified at runtime. However,
the :cache subhash is modified at runtime (safely, using a mutex).
However, Hash#hash recursively calls #hash on the values of the
hash, so this resulted in the :cache subhash being iterated over.
If concurrently, an entry is being added to the cache, this results
in a RuntimeError.

Fix this in two ways, either of which would be enough by itself:

* Precompute a hash value for AssociationReflection#hash to use,
  based on the current class and association name, and store that
  value as part of the reflection.

* Make the temporary hash a compare_by_identity hash, so it doesn't
  even trigger this issue.  It's also faster.

Even though the compare_by_identity fix is simpler, I choose to do
both of these, as it will fix issues if an association reflection
is used as a key in a user-created hash.