Make module temporary names on Ruby 3.5 consistant with 3.3 and 3.4
Proc#source_location returns starting column and ending line and
column on Ruby 3.5.  Since it seems unlikely that multiple procs
used to define modules will be used on the same line, I don't
think the extra information is useful to include in the temporary
name, and it is easier to ignore the information than change
the related specs.  However, if users have an actual need for
the additional information to disambiguate, I can revert this and
change the specs.