Add Dataset#prepare_sql_type for setting the SQL type to use for prepare/call
This allows you to separate SQL creation type from how results are
treated.  It's most useful when you want to create SQL with
INSERT/UPDATE/DELETE RETURNING, and being able to use each,
single_value, first, map, to_hash, or to_hash_groups to control
how to handle the result.  For example:

  ps = DB[:table].
    where(name: :$name).
    returning.
    prepare_sql_type(:update).
    prepare([:to_hash, :id], :update_table, value: :$value)

  # Update all rows in table with name Foo to have value Bar,
  # and return an identity-mapped hash (keys are id column values,
  # values are a hash for the related row).
  ps.call(name: "Foo", value: "Bar")

To ease implementation have Dataset#run for prepared statements use
the with_sql_* methods in most cases, so that the prepared_sql is
always provided.  In cases where with_sql_* will not work, use a
new force_prepared_sql, which will return a clone with the
prepared_sql set, if the prepared_sql_type is different from the
prepared_type.

While here, fix a case in the specs that was trying to test
prepared statements, but was testing bound variables.