Add Sequel::OLD and Sequel::NEW constants, for use in RETURNING statements on PostgreSQL 18+
PostgreSQL 18 adds support for RETURNING clauses to use OLD.column
for the old column values and NEW.column for the new column values.
The OLD and NEW cannot be quoted, because then they will be treated
as normal identifiers instead of keywords.  To make using this
feature easier on Sequel users, this adds Sequel::OLD and
Sequel::NEW.  These are SQL::Identifier instances, which can be used
to make qualified identifiers:

  Sequel::OLD[:column_name] # OLD."column_name"
  Sequel::NEW[:column_name] # NEW."column_name"