Refactor MariaDB CI support
Use similar approach used for sequel_pg, rotating not only by Ruby
minor version, but also by day of year, so that on one day, we
test MariaDB on some Ruby versions, and MySQL and others, and on
the next day, we swap, ensuring that MariaDB and MySQL are tested
on all Ruby versions on a regular basis.

Instead of having the CI set environment variables, and testing
them, assume spec_ci will only be executed in CI, and remove the
conditionals.