Consider all unique constraints when marking columns as non-NULL on DB2
DB2 requires that columns in unique constraints must be non-NULL.
Sequel has had code for a long time to automatically mark columns
as NOT NULL if they are in a unique constraint, but it apparently
only considered a single unique constraint and not multiple
unique constraints.  To avoid issues with large tables, switch
to storing the key columns in a hash.