Support :primary_key and :unique column options accepting hash value for constraint options
We previous had :primary_key_constraint_name and
:primary_key_deferrable options, but those are already a bit awkward,
and I plan on supporting more options.

This isn't done for foreign key constraints, because those do not
have a separate option name to enable them, and column options are
not separated from constraint options for foreign key constraints.
It probably would have been better to separate the constraint options
from the column options, but it's a bit too late.  Combining the
options is only problematic when a constraint option would overlap
with a column option. The only option that currently does is the
:name option, and that's handled with the :foreign_key_constaint_name
option.

The support for automatically moving named column constraints into
table constraints needs to handle this as well.  Test that it handles
additional constraint options beside the name.  It appears this
support did not handle :primary_key_deferrable or :unique_deferrable
options.  As the recommendation going forward is to provide that
as the :deferrable option inside the :primary_key/:unique option hash,
I don't plan on fixing this unless there is a user request to do so.